chore: move mentions to service hooks
